requirements.txt
hidapi>=0.14.0
psutil>=5.9.0
PyQt6>=6.6.0
